requirements.txt
duckdb>=0.9.0
streamlit>=1.30.0
altair>=5.0.0
sqlglot>=20.0.0
pyyaml>=6.0
pytest>=7.0.0
